--- react/game/src/app/screens/Login/components/fields/styles.scss
@@ -1,127 +1,0 @@
-input:not([type='checkbox']) {
-  border-width: 0;
-  border-bottom: 3px solid black;
-  border-radius: 0;
-}
-
-input:not([type='checkbox']) + label {
-  transition: all 0.2s;
-}
-
-input:not([type='checkbox']) + label {
-  position: absolute;
-  left: 0;
-  top: 0;
-  pointer-events: none;
-  color: gray;
-}
-
-input:not([type='checkbox']):focus + label,
-.dirty input:not([type='checkbox']) + label {
-  font-size: 12px;
-  top: -12px;
-  color: black;
-}
-
-.flex-row-reverse {
-  display: flex;
-  justify-content: space-between;
-  flex-direction: row-reverse;
-}
-
-input {
-  outline: 0;
-  padding: 0;
-}
-
-input,
-label,
-button[type='submit'] {
-  font-size: 20px;
-}
-
-select {
-  height: 30px;
-}
-
-label {
-  font-size: 12px;
-}
-
-.custom-input-container,
-.custom-select-container,
-.custom-field-array-container {
-  position: relative;
-  margin: 0 0 25px 0;
-}
-
-select {
-  display: block;
-}
-
-select,
-input:not([type='checkbox']),
-button[type='submit'] {
-  width: 100%;
-}
-
-button[type='submit'] {
-  background-color: orange;
-  color: white;
-}
-
-/* Validation styles */
-.valid input {
-  border-color: green;
-}
-
-.invalid input {
-  border-color: red;
-}
-
-.error-text {
-  background-color: red;
-  color: white;
-  margin-top: 10px;
-  position: relative;
-}
-
-.error-text::before {
-  content: '';
-  border-color: transparent transparent red transparent;
-  width: 0;
-  height: 0;
-  border-style: solid;
-  border-width: 0 8px 8px 7px;
-  position: absolute;
-  left: 20px;
-  top: -8px;
-}
-
-@keyframes grey-pulse {
-  0% {
-    border-color: grey;
-  }
-  50% {
-    border-color: #eee;
-  }
-  100% {
-    border-color: grey;
-  }
-}
-
-.async-validating input {
-  border-color: gray;
-  animation-name: grey-pulse;
-  animation-iteration-count: infinite;
-  animation-duration: 1.4s;
-}
-
-.field-array-item {
-  display: flex;
-  justify-content: space-between;
-}
-
-.field-array-item button {
-  align-self: start;
-}
