--- react/game/src/app/screens/Login/components/loginForm/styles.scss
@@ -1,5 +1,0 @@
-.login-form {
-  max-width: 320px;
-  padding: 30px;
-  margin: auto;
-}
